REFAC: moved navigators into a navigator folder
