+++ App.js
@@ -9,8 +9,8 @@ import Signup from './components/Signup';
 import GameOptionPage from './components/GameOptionPage';
 import Leaderboard from './components/Leaderboard';
 import CustomGameList from './components/CustomGameList';
-import TutorialNavigator from './components/TutorialNavigator';
-import CustomTutorialNavigator from './components/CustomTutorialNavigator';
+import TutorialNavigator from './components/Navigators/TutorialNavigator';
+import CustomTutorialNavigator from './components/Navigators/CustomTutorialNavigator';
 import { isSignedIn } from './auth';
 
 const RootNavigator = StackNavigator(
